requirements.txt
tensorflow==1.14.0
Werkzeug==1.0.1
numpy==1.19.2
opencv_python==4.4.0.44
Flask==1.1.2
six==1.11.0
object_detection==0.0.3
Pillow==7.2.0
protobuf==3.13.0
